+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - update tough-cookie to remove punycode deprecation warning (#813)
 - support $env intellij environment variables access (#811)
 - failed tests in reference skips request (AnWeber/vscode-httpyac#335)
+- support `http_proxy` environment variable
 
 ## [6.15.1] ( 2024-08-22)
 ### Features